ingestion/requirements.txt
fastapi
uvicorn
google-cloud-storage
google-cloud-bigquery
google-cloud-aiplatform
pypdf
langchain-text-splitters
